refactor(ReportsRoutes): move compoent to its own directory
